Refined OPFParser. New Metadata/NCXDocument/NavigationDocument Parsers classes and files. Modularity achieved, still a few changes to make.
TODO: clean and refactor MetadataParser.
