+++ Sources/parser/MetadataParser.swift
@@ -1,0 +1,207 @@
+//
+//  MetadataParser.swift
+//  R2Streamer
+//
+//  Created by Alexandre Camilleri on 3/17/17.
+//  Copyright © 2017 Readium. All rights reserved.
+//
+
+import Foundation
+import AEXML
+
+// MARK: - Metadata Parsing.
+public class MetadataParser {
+
+    /// Extracts the Rendition properties from the XML element metadata and fill
+    /// then into the Metadata object instance.
+    ///
+    /// - Parameters:
+    ///   - metadataElement: The XML element containing the metadatas.
+    ///   - metadata: The `Metadata` object.
+    internal func setRenditionProperties(from metadataElement: AEXMLElement,
+                                         to metadata: inout Metadata) {
+        // Layout
+        var attribute = ["property" : "rendition:layout"]
+
+        if let renditionLayouts = metadataElement.all(withAttributes: attribute),
+            !renditionLayouts.isEmpty {
+            let layouts = renditionLayouts[0].string
+
+            metadata.rendition.layout = RenditionLayout(rawValue: layouts)
+        }
+        // Flow
+        attribute = ["property" : "rendition:flow"]
+        if let renditionFlows = metadataElement.all(withAttributes: attribute),
+            !renditionFlows.isEmpty {
+            let flows = renditionFlows[0].string
+
+            metadata.rendition.flow = RenditionFlow(rawValue: flows)
+        }
+        // Orientation
+        attribute = ["property" : "rendition:orientation"]
+        if let renditionOrientations = metadataElement.all(withAttributes: attribute),
+            !renditionOrientations.isEmpty {
+            let orientation = renditionOrientations[0].string
+
+            metadata.rendition.orientation = RenditionOrientation(rawValue: orientation)
+        }
+        // Spread
+        attribute = ["property" : "rendition:spread"]
+        if let renditionSpreads = metadataElement.all(withAttributes: attribute),
+            !renditionSpreads.isEmpty {
+            let spread = renditionSpreads[0].string
+
+            metadata.rendition.spread = RenditionSpread(rawValue: spread)
+        }
+        // Viewport
+        attribute = ["property" : "rendition:viewport"]
+        if let renditionViewports = metadataElement.all(withAttributes: attribute),
+            !renditionViewports.isEmpty {
+            metadata.rendition.viewport = renditionViewports[0].string
+        }
+    }
+
+    /// Get the main title of the publication from the from the OPF XML document
+    /// `<metadata>` element.
+    ///
+    /// - Parameter metadata: The `<metadata>` element.
+    /// - Returns: The content of the `<dc:title>` element, `nil` if the element
+    ///            wasn't found.
+    internal func parseMainTitle(from metadata: AEXMLElement, epubVersion: Double?) -> String? {
+        // Return if there isn't any `<dc:title>` element
+        guard let titles = metadata["dc:title"].all else {
+            return nil
+        }
+        // If there's more than one, look for the `main` one as defined by
+        // `refines`.
+        // Else, as a fallback and default, return the first `<dc:title>`
+        // content.
+        guard titles.count > 1, epubVersion == 3 else {
+            return metadata["dc:title"].string
+        }
+        /// Used in the closure below.
+        func isMainTitle(element: AEXMLElement) -> Bool {
+            guard let eid = element.attributes["id"] else {
+                return false
+            }
+            let attributes = ["property": "title-type", "refines": "#" + eid]
+            let metas = metadata["meta"].all(withAttributes: attributes)
+
+            return metas?.contains(where: { $0.string == "main" }) ?? false
+        }
+        // Returns the first main title encountered
+        return titles.first(where: { isMainTitle(element: $0)})?.string
+    }
+
+    /// Get the unique identifer of the publication from the from the OPF XML
+    /// document `<metadata>` element.
+    ///
+    /// - Parameters:
+    ///   - metadata: The `<metadata>` element.
+    ///   - Attributes: The XML document attributes.
+    /// - Returns: The content of the `<dc:identifier>` element, `nil` if the
+    ///             element wasn't found.
+    internal func parseUniqueIdentifier(from metadata: AEXMLElement,
+                                        withAttributes attributes: [String : String]) -> String? {
+        // Look for `<dc:identifier>` elements
+        guard let identifiers = metadata["dc:identifier"].all else {
+            return nil
+        }
+        // Get the one defined as unique by the `<package>` attribute
+        // `unique-identifier`
+        if identifiers.count > 1,
+            let uniqueId = attributes["unique-identifier"] {
+            let uniqueIdentifiers = identifiers.filter { $0.attributes["id"] == uniqueId }
+            if !uniqueIdentifiers.isEmpty, let uid = uniqueIdentifiers.first {
+                return uid.string
+            }
+        }
+        // Returns the first `<dc:identifier>` content or an empty String
+        return metadata["dc:identifier"].string
+    }
+
+    /// Builds a `Contributor` instance from a `<dc:creator>` or
+    /// `<dc:contributor>` element.
+    ///
+    /// - Parameters:
+    ///   - element: The XML element to parse.
+    ///   - doc: The OPF XML document being parsed (necessary to look for
+    ///          `refines`).
+    /// - Returns: The contributor instance filled with its name and optionally
+    ///            its `role` and `sortAs` attributes.
+    internal func createContributor(from element: AEXMLElement, metadata: AEXMLElement,
+                                    epubVersion: Double?) -> Contributor {
+        // The 'to be returned' Contributor object.
+        let contributor = Contributor(name: element.string)
+
+        // Get role from role attribute
+        if let role = element.attributes["opf:role"] {
+            contributor.role = role
+        }
+        // Get sort name from file-as attribute
+        if let sortAs = element.attributes["opf:file-as"] {
+            contributor.sortAs = sortAs
+        }
+        // Look up for possible meta refines for role
+        if epubVersion == 3, let eid = element.attributes["id"] {
+            let attributes = ["property": "role", "refines": "#\(eid)"]
+
+            if let metas = metadata["meta"].all(withAttributes: attributes),
+                !metas.isEmpty, let first = metas.first {
+                let role = first.string
+
+                contributor.role = role
+            }
+        }
+        return contributor
+    }
+
+    /// Parse a `creator` or `contributor` element from the OPF XML document,
+    /// then builds and adds a Contributor to the metadata, to an array
+    /// according to its role (authors, translators, etc.).
+    ///
+    /// - Parameters:
+    ///   - element: The XML element to parse.
+    ///   - doc: The OPF XML document being parsed.
+    ///   - metadata: The metadata to which to add the contributor.
+    internal func parseContributor(from element: AEXMLElement,
+                                   in document: AEXMLDocument,
+                                   to metadata: Metadata,
+                                   with epubVersion: Double?) {
+        let metadataElement = document.root["metadata"]
+        let contributor = createContributor(from: element,
+                                            metadata: metadataElement,
+                                            epubVersion: epubVersion)
+
+        // Add the contributor to the proper property according to the its `role`
+        if let role = contributor.role {
+            switch role {
+            case "aut":
+                metadata.authors.append(contributor)
+            case "trl":
+                metadata.translators.append(contributor)
+            case "art":
+                metadata.artists.append(contributor)
+            case "edt":
+                metadata.editors.append(contributor)
+            case "ill":
+                metadata.illustrators.append(contributor)
+            case "clr":
+                metadata.colorists.append(contributor)
+            case "nrt":
+                metadata.narrators.append(contributor)
+            case "pbl":
+                metadata.publishers.append(contributor)
+            default:
+                metadata.contributors.append(contributor)
+            }
+        } else {
+            // No role, so add the creators to the authors and the others to the contributors
+            if element.name == "dc:creator" {
+                metadata.authors.append(contributor)
+            } else {
+                metadata.contributors.append(contributor)
+            }
+        }
+    }
+}
